Added continuous integration
This patch refurbishes the code to set up the continuous integration on
GitHub Workflows. In particular, the component test was refactored into
multiple unit tests which required a couple of changes to the interface
so that the code can be tested.

Additionally, while we are at it, we add support for Python 3.6, 3.7 and
3.8 as well as update mypy to 0.790 and pylint to 2.6.0.